+++ src/main/java/com/caskey/apibuilder/service/GetService.java
@@ -8,7 +8,7 @@ public interface GetService<T extends BaseEntity, D extends BaseEntityDTO> exten
         return getRepository().findOne(id);
     }
 
-    default D getDTOById(final Long id, final Long depth) {
+    default D getDTOById(final Long id, final Integer depth) {
         return getAdapter().toDTO(getById(id), depth);
     }
 }
